+++ src/Carbunql.TypeSafe/FluentSelectQuery.cs
@@ -136,50 +136,16 @@ public class FluentSelectQuery : SelectQuery
         }
         else if (exp is ConditionalExpression cnd)
         {
-            return CreateCaseStatement(cnd, addParameter);
+            return cnd.ToValue(ToValue, addParameter);
         }
 
         throw new InvalidProgramException(exp.ToString());
     }
 
-
     internal static string CreateCastStatement(string value, Type type)
     {
         var dbtype = DbmsConfiguration.ToDbType(type);
         return $"cast({value} as {dbtype})";
-    }
-
-    private string CreateCaseStatement(ConditionalExpression cnd, Func<object?, string> addParameter)
-    {
-        var test = ToValue(cnd.Test, addParameter);
-        var ifTrue = ToValue(cnd.IfTrue, addParameter);
-        var ifFalse = ToValue(cnd.IfFalse, addParameter);
-
-        if (string.IsNullOrEmpty(ifFalse))
-        {
-            throw new ArgumentException("The IfFalse expression cannot be null or empty.", nameof(cnd.IfFalse));
-        }
-
-        // When case statements are nested, check if there is an alternative in the when clause
-        if (ifFalse.TrimStart().StartsWith("case ", StringComparison.OrdinalIgnoreCase))
-        {
-            var caseExpression = CaseExpressionParser.Parse(ifFalse);
-            if (caseExpression.CaseCondition is null)
-            {
-                // Replace with when clause
-                var we = WhenExpressionParser.Parse($"when {test} then {ifTrue}");
-                caseExpression.WhenExpressions.Insert(0, we);
-                return caseExpression.ToText();
-            }
-            else
-            {
-                return $"case when {test} then {ifTrue} else {ifFalse} end";
-            }
-        }
-        else
-        {
-            return $"case when {test} then {ifTrue} else {ifFalse} end";
-        }
     }
 
     internal static string ConverToDbDateFormat(string csharpFormat)
